Do no reevaluate local components
In the past we permitted mutations of a configuration after the component state for that configurations
root component was generated. Now, we no longer need to do so as we do not permit mutating a configuration
after it has been observed. Now, local component state instances can be functionally immutable and no longer
need to expose the reevaluate functionality