Add back reevaluate functionality
We remove this in hope that it was no longer necessary. However, without this logic it is possible
for resolutions in a project, which consume their own variants, to see two different versions of a
set of dependencies.

This is apparent when a resolvable configuration and consumable configuration extend the same dependency
scope, and that dependency scope is modified inside the beforeResolve callback of the resolvable
configuration.

This behavior currently emits a deprecation warning. When we make that warning into an error, we can remove
this reevaluate code.